split counter and gauge into their own files
