+++ staging_vespalib/src/vespa/vespalib/metrics/mergers.h
@@ -4,6 +4,8 @@
 #include <mutex>
 #include "no_realloc_bunch.h"
 #include "simple_metrics.h"
+#include "counter.h"
+#include "gauge.h"
 #include "clock.h"
 
 namespace vespalib {
